% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nused_data.R
\name{get_nused_data}
\alias{get_nused_data}
\title{Reads NuSEDS Data}
\usage{
get_nused_data()
}
\value{
tibble with NuSEDS salmon data exported in 2018.
}
\description{
Reads data that was exported from the New Salmon Escapement Database System (NuSEDS).
}
\details{
Reads from a 136 MB csv file of New Salmon Escapement Database System (NuSEDS) data for 2018.
   \itemize{
     \item data: http://www.pac.dfo-mpo.gc.ca/od-ds/science/sed-des/NUSEDS_20180416.zip
     \item metadata: https://open.canada.ca/data/en/dataset/c48669a3-045b-400d-b730-48aafe8c5ee6
   }
}
