% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_data.R
\name{plot_single_fpc_file}
\alias{plot_single_fpc_file}
\title{Plots FPC Data for a Single Year and Single Passage Reporting Site.}
\usage{
plot_single_fpc_file(x, year, passage_reporting_site)
}
\arguments{
\item{x}{data.frame with one year of FPC observations.}

\item{year}{integer year in which the observations were made.
This value appears in the title of the plot.}

\item{passage_reporting_site}{string name of observation point.
For example, "Bonneville Dam".
This value appears in the title of the plot.}
}
\value{
plot object from ggplot with quantity of daily
    observations for species: Chinook, Coho, Sockeye, Steelhead
}
\description{
Plots a single year and single passage reporting site of
    Fish Passage Center (FPC) data contained
    in a single file with only that year of data.
}
